+++ R/ncurl.R
@@ -27,12 +27,15 @@
 #' @param follow [default FALSE] logical value whether to automatically follow
 #'     redirects (not applicable for async requests). If FALSE, the redirect
 #'     address is returned as response header 'Location'.
-#' @param method (optional) the HTTP method (defaults to 'GET' if not specified).
+#' @param method (optional) the HTTP method as a character string. Defaults to
+#'     'GET' if not specified, and could also be 'POST', 'PUT' etc.
 #' @param headers (optional) a named character vector specifying the HTTP
 #'     request headers, for example: \cr
 #'     \code{c(Authorization = "Bearer APIKEY", `Content-Type` = "text/plain")}
 #'     \cr A non-character or non-named vector will be ignored.
-#' @param data (optional) character request data to be submitted.
+#' @param data (optional) character string request data to be submitted. If a
+#'     vector, only the first element is taken, and non-character objects are
+#'     ignored.
 #' @param response (optional) a character vector specifying the response headers
 #'     to return e.g. \code{c("date", "server")}. These are case-insensitive and
 #'     will return NULL if not present. A non-character vector will be ignored.
@@ -122,7 +125,7 @@ ncurl <- function(url,
 #'
 #' @seealso \code{\link{ncurl_session}} for persistent connections.
 #' @examples
-#' nc <- ncurl_aio("https://www.r-project.org/",
+#' nc <- ncurl_aio("https://postman-echo.com/get",
 #'                 response = c("date", "server"),
 #'                 timeout = 2000L)
 #' call_aio(nc)
@@ -132,6 +135,7 @@ ncurl <- function(url,
 #'
 #' if (interactive() && requireNamespace("promises", quietly = TRUE)) {
 #'
+#' library(promises)
 #' p <- as.promise(nc)
 #' print(p)
 #'
@@ -168,10 +172,12 @@ ncurl_aio <- function(url,
 #'
 #' @seealso \code{\link{ncurl_aio}} for asynchronous http requests.
 #' @examples
-#' s <- ncurl_session("https://www.r-project.org/", response = "date", timeout = 2000L)
+#' s <- ncurl_session("https://postman-echo.com/get",
+#'                    response = "date",
+#'                    timeout = 2000L)
 #' s
-#' if (!is_error_value(s)) transact(s)
-#' if (!is_error_value(s)) close(s)
+#' if (is_ncurl_session(s)) transact(s)
+#' if (is_ncurl_session(s)) close(s)
 #'
 #' @export
 #'