+++ R/opts.R
@@ -46,6 +46,20 @@
 #'     For an authoritative guide please refer to the online documentation for
 #'     the NNG library at \url{https://nng.nanomsg.org/man/}.
 #'
+#' @section Serialization:
+#'
+#'     Apart from the NNG options documented below, there is the following
+#'     special option:
+#'
+#'     \itemize{
+#'       \item 'serial' [type list]
+#'
+#'       For Sockets only. This accepts a configuration created by
+#'       \code{\link{serial_config}}. Setting a new configuration replaces any
+#'       already set. To remove entirely, supply an empty list. Note: this
+#'       option is write-only and can be set but not retrieved.
+#'     }
+#'
 #' @section Global Options:
 #'
 #'     \itemize{
@@ -113,6 +127,21 @@
 #'       be queued for delivery by the socket for this period of time (such as
 #'       if send buffers are full), the operation will fail with a return value
 #'       of 5L ('timed out').
+#'
+#'     \item 'recv-fd' [type 'int']
+#'
+#'       This is the socket receive file descriptor. For supported protocols,
+#'       this will become readable when a message is available for receiving on
+#'       the socket. Attempts should not be made to read or write to the
+#'       returned file descriptor, but it is suitable for use with poll(),
+#'       select(), or WSAPoll() on Windows, and similar functions.
+#'
+#'     \item 'send-fd' [type 'int']
+#'
+#'       This is the socket send file descriptor. Attempts should not be made to
+#'       read or write to the returned file descriptor, but it is suitable for
+#'       use with poll(), select(), or WSAPoll() on Windows, and similar
+#'       functions.
 #'
 #'     \item 'socket-name' [type 'string']
 #'
@@ -299,7 +328,7 @@ opt <- function(object, name)
 #'     pattern. Set a topic to subscribe to, or remove a topic from the
 #'     subscription list.
 #'
-#' @param con a Socket or Context using the 'sub' protocol.
+#' @param con a Socket or Context using the \sQuote{sub} protocol.
 #' @param topic [default NULL] an atomic type or NULL. The default NULL
 #'     subscribes to all topics / unsubscribes from all topics (if all topics
 #'     were previously subscribed).
@@ -357,7 +386,7 @@ unsubscribe <- function(con, topic = NULL)
 #'     subsequent surveys). Messages received by the surveyor after the timer
 #'     has ended are discarded.
 #'
-#' @param con a Socket or Context using the 'surveyor' protocol.
+#' @param con a Socket or Context using the \sQuote{surveyor} protocol.
 #' @param value [default 1000L] integer survey timeout in milliseconds.
 #'
 #' @return Invisibly, the passed Socket or Context.