+++ src/nanonext.h
@@ -20,10 +20,13 @@
 #define NANONEXT_H
 
 #include <nng/nng.h>
+#include <nng/supplemental/util/platform.h>
+#include <nng/supplemental/tls/tls.h>
 
 #ifdef NANONEXT_PROTOCOLS
 #include <nng/protocol/bus0/bus.h>
 #include <nng/protocol/pair0/pair.h>
+#include <nng/protocol/pair1/pair.h>
 #include <nng/protocol/pubsub0/pub.h>
 #include <nng/protocol/pubsub0/sub.h>
 #include <nng/protocol/pipeline0/pull.h>
@@ -34,64 +37,8 @@
 #include <nng/protocol/survey0/respond.h>
 #endif
 
-#ifdef NANONEXT_SUPPLEMENTALS
+#ifdef NANONEXT_HTTP
 #include <nng/supplemental/http/http.h>
-#include <nng/supplemental/tls/tls.h>
-#include <nng/supplemental/util/platform.h>
-
-typedef union nano_opt_u {
-  char *str;
-  bool b;
-  nng_duration d;
-  int i;
-  size_t s;
-  uint64_t u;
-} nano_opt;
-
-typedef struct nano_listener_s {
-  nng_listener list;
-  nng_tls_config *tls;
-} nano_listener;
-
-typedef struct nano_dialer_s {
-  nng_dialer dial;
-  nng_tls_config *tls;
-} nano_dialer;
-
-typedef struct nano_stream_s {
-  nng_stream *stream;
-  enum {
-    NANO_STREAM_DIALER,
-    NANO_STREAM_LISTENER
-  } mode;
-  int textframes;
-  union {
-    nng_stream_dialer *dial;
-    nng_stream_listener *list;
-  } endpoint;
-  nng_tls_config *tls;
-} nano_stream;
-
-typedef enum nano_aio_typ {
-  SENDAIO,
-  RECVAIO,
-  REQAIO,
-  IOV_SENDAIO,
-  IOV_RECVAIO,
-  HTTP_AIO,
-  RECVAIOS,
-  REQAIOS,
-  IOV_RECVAIOS
-} nano_aio_typ;
-
-typedef struct nano_aio_s {
-  nng_aio *aio;
-  nano_aio_typ type;
-  int mode;
-  int result;
-  void *data;
-  void *next;
-} nano_aio;
 
 typedef struct nano_handle_s {
   nng_url *url;
@@ -100,30 +47,6 @@ typedef struct nano_handle_s {
   nng_http_res *res;
   nng_tls_config *cfg;
 } nano_handle;
-
-typedef struct nano_cv_s {
-  int condition;
-  int flag;
-  nng_mtx *mtx;
-  nng_cv *cv;
-} nano_cv;
-
-typedef struct nano_cv_duo_s {
-  nano_cv *cv;
-  nano_cv *cv2;
-} nano_cv_duo;
-
-typedef struct nano_thread_aio_s {
-  nng_thread *thr;
-  nano_cv *cv;
-  nng_aio *aio;
-} nano_thread_aio;
-
-typedef struct nano_thread_duo_s {
-  nng_thread *thr;
-  nano_cv *cv;
-  nano_cv *cv2;
-} nano_thread_duo;
 
 #endif
 
@@ -142,26 +65,6 @@ typedef struct nano_thread_duo_s {
 #endif
 
 #ifdef NANONEXT_TLS
-#include <mbedtls/base64.h>
-#include <mbedtls/md.h>
-#include <mbedtls/sha256.h>
-#include <mbedtls/sha512.h>
-#include <mbedtls/version.h>
-
-#define SHA224_KEY_SIZE 28
-#define SHA256_KEY_SIZE 32
-#define SHA384_KEY_SIZE 48
-#define SHA512_KEY_SIZE 64
-
-#endif
-
-#ifdef NANONEXT_MBED
-#include <mbedtls/entropy.h>
-#include <mbedtls/ctr_drbg.h>
-#include <mbedtls/error.h>
-#endif
-
-#ifdef NANONEXT_KEYCERT
 #include <mbedtls/version.h>
 #if MBEDTLS_VERSION_MAJOR < 3
 #include <mbedtls/config.h>
@@ -192,6 +95,8 @@ typedef struct nano_thread_duo_s {
 #define NANO_SET_PROT(x, v) SETCDR(x, v)
 #define NANO_DATAPTR(x) (void *) DATAPTR_RO(x)
 #define NANO_VECTOR(x) ((const SEXP *) DATAPTR_RO(x))
+#define NANO_STRING(x) CHAR(*((const SEXP *) DATAPTR_RO(x)))
+#define NANO_STR_N(x, n) CHAR(((const SEXP *) DATAPTR_RO(x))[n])
 #define NANO_INTEGER(x) *(int *) DATAPTR_RO(x)
 #define NANO_ERROR(x) { Rf_error(x); return R_NilValue; }
 
@@ -216,37 +121,117 @@ typedef struct nano_thread_duo_s {
   Rf_classgets(x, klass);                                      \
   SET_STRING_ELT(klass, 0, Rf_mkChar(cls1));                   \
   SET_STRING_ELT(klass, 1, Rf_mkChar(cls2))
+
+typedef union nano_opt_u {
+  char *str;
+  bool b;
+  nng_duration d;
+  int i;
+  size_t s;
+  uint64_t u;
+} nano_opt;
+
+typedef struct nano_listener_s {
+  nng_listener list;
+  nng_tls_config *tls;
+} nano_listener;
+
+typedef struct nano_dialer_s {
+  nng_dialer dial;
+  nng_tls_config *tls;
+} nano_dialer;
+
+typedef struct nano_stream_s {
+  nng_stream *stream;
+  union {
+    nng_stream_dialer *dial;
+    nng_stream_listener *list;
+  } endpoint;
+  nng_tls_config *tls;
+  int textframes;
+  enum {
+    NANO_STREAM_DIALER,
+    NANO_STREAM_LISTENER
+  } mode;
+} nano_stream;
+
+typedef enum nano_aio_typ {
+  SENDAIO,
+  RECVAIO,
+  REQAIO,
+  IOV_SENDAIO,
+  IOV_RECVAIO,
+  HTTP_AIO,
+  RECVAIOS,
+  REQAIOS,
+  IOV_RECVAIOS
+} nano_aio_typ;
+
+typedef struct nano_aio_s {
+  nng_aio *aio;
+  void *data;
+  void *next;
+  int mode;
+  int result;
+  nano_aio_typ type;
+} nano_aio;
+
+typedef struct nano_cv_s {
+  int condition;
+  int flag;
+  nng_mtx *mtx;
+  nng_cv *cv;
+} nano_cv;
+
+typedef struct nano_cv_duo_s {
+  nano_cv *cv;
+  nano_cv *cv2;
+} nano_cv_duo;
+
+typedef struct nano_thread_aio_s {
+  nng_thread *thr;
+  nano_cv *cv;
+  nng_aio *aio;
+} nano_thread_aio;
+
+typedef struct nano_thread_duo_s {
+  nng_thread *thr;
+  nano_cv *cv;
+  nano_cv *cv2;
+} nano_thread_duo;
+
 typedef struct nano_buf_s {
   unsigned char *buf;
   size_t len;
   size_t cur;
 } nano_buf;
 
+void dialer_finalizer(SEXP);
+void listener_finalizer(SEXP);
+void socket_finalizer(SEXP);
+SEXP R_mkClosure(SEXP, SEXP, SEXP);
 void later2(void (*)(void *), void *);
 extern void (*eln2)(void (*)(void *), void *, double, int);
 void eln2dummy(void (*)(void *), void *, double, int);
-
 int nano_integer(SEXP);
-SEXP nano_PreserveObject(SEXP);
-void nano_ReleaseObject(SEXP);
 SEXP mk_error(const int);
-nano_buf nano_char_buf(const SEXP);
-SEXP nano_decode(unsigned char *, const size_t, const int);
+SEXP mk_error_data(const int);
+SEXP rawToChar(const unsigned char *, const size_t);
+void nano_serialize_old(nano_buf *, const SEXP, SEXP);
+void nano_serialize(nano_buf *, const SEXP, SEXP);
+SEXP nano_unserialize(unsigned char *, const size_t, SEXP);
+SEXP nano_decode(unsigned char *, const size_t, const int, SEXP);
 void nano_encode(nano_buf *, const SEXP);
 int nano_encodes(const SEXP);
 int nano_matcharg(const SEXP);
 int nano_matchargs(const SEXP);
-void nano_serialize(nano_buf *, const SEXP);
-void nano_serialize_next(nano_buf *, const SEXP);
-void nano_serialize_xdr(nano_buf *, const SEXP);
-SEXP nano_unserialize(unsigned char *, const size_t);
-SEXP rawToChar(const unsigned char *, const size_t);
-void dialer_finalizer(SEXP);
-void listener_finalizer(SEXP);
-void socket_finalizer(SEXP);
+SEXP nano_PreserveObject(SEXP);
+void nano_ReleaseObject(SEXP);
 
+SEXP rnng_advance_rng_state(void);
 SEXP rnng_aio_call(SEXP);
 SEXP rnng_aio_collect(SEXP);
+SEXP rnng_aio_collect_pipe(SEXP);
 SEXP rnng_aio_collect_safe(SEXP);
 SEXP rnng_aio_get_msg(SEXP);
 SEXP rnng_aio_http_data(SEXP);
@@ -285,8 +270,9 @@ SEXP rnng_ncurl_session(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP rnng_ncurl_session_close(SEXP);
 SEXP rnng_ncurl_transact(SEXP);
 SEXP rnng_next_config(SEXP, SEXP, SEXP, SEXP);
+SEXP rnng_pipe_close(SEXP);
 SEXP rnng_pipe_notify(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-SEXP rnng_protocol_open(SEXP, SEXP);
+SEXP rnng_protocol_open(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP rnng_random(SEXP, SEXP);
 SEXP rnng_reap(SEXP);
 SEXP rnng_recv(SEXP, SEXP, SEXP, SEXP);
@@ -294,6 +280,8 @@ SEXP rnng_recv_aio(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP rnng_request(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP rnng_send(SEXP, SEXP, SEXP, SEXP);
 SEXP rnng_send_aio(SEXP, SEXP, SEXP, SEXP, SEXP);
+SEXP rnng_serial_config(SEXP, SEXP, SEXP, SEXP);
+SEXP rnng_set_marker(SEXP);
 SEXP rnng_set_promise_context(SEXP, SEXP);
 SEXP rnng_set_opt(SEXP, SEXP, SEXP);
 SEXP rnng_signal_thread_create(SEXP, SEXP);
@@ -315,6 +303,8 @@ SEXP rnng_version(void);
 SEXP rnng_wait_thread_create(SEXP);
 SEXP rnng_write_cert(SEXP, SEXP, SEXP);
 
+extern uint8_t special_bit;
+
 extern SEXP nano_AioSymbol;
 extern SEXP nano_ContextSymbol;
 extern SEXP nano_CvSymbol;
@@ -324,8 +314,8 @@ extern SEXP nano_DotcallSymbol;
 extern SEXP nano_HeadersSymbol;
 extern SEXP nano_IdSymbol;
 extern SEXP nano_ListenerSymbol;
+extern SEXP nano_PipeSymbol;
 extern SEXP nano_ProtocolSymbol;
-extern SEXP nano_RawSymbol;
 extern SEXP nano_ResolveSymbol;
 extern SEXP nano_ResponseSymbol;
 extern SEXP nano_ResultSymbol;
@@ -345,7 +335,6 @@ extern SEXP nano_klassString;
 extern SEXP nano_precious;
 extern SEXP nano_recvAio;
 extern SEXP nano_reqAio;
-extern SEXP nano_refHook;
 extern SEXP nano_sendAio;
 extern SEXP nano_success;
 extern SEXP nano_unresolved;