+++ NEWS.md
@@ -1,3 +1,23 @@
+# nanonext 1.2.0
+
+#### New Features
+
+* Adds `serial_config()` to create configurations that can be set on Sockets to make use of custom serialization and unserialization functions for reference objects (plugs into the 'refhook' system of native R serialization).
+* `'opt<-'()` now accepts the special option 'serial' for Sockets, which takes a configuration returned from `serial_config()`.
+* Adds the 'poly' protocol for one-to-one of many socket connections (NNG's pair v1 polyamorous mode).
+* Adds `is_ncurl_session()` as a validation function.
+* Adds `collect_pipe()` for obtaining the underlying Pipe from a 'recvAio'. This affords more granular control of connections, with the ability to close individual pipes.
+* `send_aio()` now accept a Pipe to direct messages to a specific peer for supported protocols such as 'poly'.
+
+#### Updates
+
+* Send mode 'next' is folded into the default 'serial', with custom serialization functions applying automatically if they have been registered.
+* The session-wide `next_config()` is now deprecated and defunct, in favour of the new `serial_config()`.
+* `ncurl_session()` now returns 'errorValue' 7 (Object closed) when attempting to transact over a closed session or closing a closed session, rather than throwing an error.
+* `collect_aio()` and `collect_aio_()` no longer append empty names when acting on lists of Aios where there were none in the first place.
+* Removes hard dependency on `stats` and `utils` base packages.
+* Requires R >= 3.6.
+
 # nanonext 1.1.1
 
 #### New Features